+++ srai/loaders/__init__.py
@@ -2,5 +2,6 @@
 
 from .base import BaseLoader
 from .geoparquet_loader import GeoparquetLoader
+from .gtfs_loader import GTFSLoader
 
-__all__ = ["BaseLoader", "GeoparquetLoader"]
+__all__ = ["BaseLoader", "GeoparquetLoader", "GTFSLoader"]